Migrate websocket-transport
